feat: Define per distro/profile/system snippets
Allow to define a new key `snippets` in `cobbler_distros`,
`cobbler_profiles` and `cobbler_systems`.

This is a list of dictionaries with the name of the snippet and its
content.

Example:
```
cobbler_profiles:
  - name: profile_name
    snippets:
      - name: partition
        content: |
          autopart --type=lvm
```

This will create the file
`/var/lib/cobbler/snippets/per_profile/partition/profile_name` with the
content `autopart --type=lvm`.