feat: add a new snippet custom_post_install
This will install a new snippet that will be executed during the post
installation process, in the script preseed_late_default. By default,
the snippet is empty. One can define a custom script with the variable
`cobbler_custom_post_install`.

It is also possible to override this snippet with a per profile, distro
or system snippet.