feat: Add autoinstall meta variable packages
The new autoinstall meta variable `packages` allows to install
additional packages during the provisioning of the system.

For RHEL-like, inject the packages in the template file. The list of
packages should include one package per line. For this, define the list
of packages with YAML block scalar style literal (`|`).

```
  - name: profileA
    properties:
      distro: almalinux-8
      autoinstall_meta:
        packages: |
          cloud-init
          qemu-guest-agent
```

For Debian-like, it's a bit trickier to add packages to a single line,
considering conditions already exists to select a list of base packages.
Instead, install the packages with apt during post-installation stage.

The list of packages should be provided as a single line. For this,
define the list with YAML block scalar style folded (`>`).

```
  - name: profileB
    properties:
      distro: debian-11
      autoinstall_meta:
        packages: >
          cloud-init
          qemu-guest-agent
```

Note: If one should define their own script with `custom_post_install`,
they would have to import the logic to install additional package in
their new script.